Frequency Calculations Completed
I believe I have completed the .cpp file, though I have not tested it yet. I have also begun the main file. I'm not sure by the way the README is written but, they may want us to output the frequency map to a file named 'frequencies'. If so, that can be easily implemented.